Fix pagination for assets
We weren't considering the ignored assets handling logic when returning
the total number of entries found. Since information is split between
the userdb and the globaldb we need in some cases to iterate manually
over the assets.
